// src/app/models/colies.ts
export interface Colies {

    colieId?: string;
    nameLot?: string;
    agencyId: string;
    lotId: string;
    cinSender: string;
    firstNameSender: string;
    lastNameSender: string;
    streetSender: string;
    mobileSender: string;
    emailSender: string;
    cinRecipient: string;
    firstNameRecipient: string;
    lastNameRecipient: string;
    countryRecipient: string;
    cityRecipient: string;
    streetRecipient: string;
    mobileRecipient: string;
    emailRecipient: string;
    delivred?: string;

}